requirements.txt
kafka-python==2.0.2
yfinance==0.2.44
pandas==2.2.2
numpy==2.0.1
requests==2.32.3
vaderSentiment==3.3.2
pyspark==3.5.1
redis==5.0.7
dash==2.17.1
plotly==5.23.0
